Add BlueyOS cross-compilation support and related configurations
- Introduced `configure-blueyos` script for easier cross-compilation setup with the BlueyOS toolchain.
- Added auto-detection logic for the BlueyOS cross-compiler and sysroot in `configure.ac`.
- Created `blueyos-cc` wrapper to handle GCC specs for cross-compilation.
- Generated `blueyos-cross.specs` to specify include paths and startup files for the BlueyOS environment.
- Updated `MUSL-PATHS.md` documentation to reflect new build instructions and auto-detection conditions.
- Modified `paths.c` to handle prefix length checks with proper formatting.
- Refactored `supervisor.c` to replace `vfork()` with `fork()` for safer process creation.
- Added `claw-build.h` for build metadata.

+++ src/core/config/paths.c
@@ -39,8 +39,8 @@ const struct claw_paths *claw_get_paths(void) {
     size_t plen = strlen(raw_prefix);
     if (plen >= sizeof(prefix_buf)) {
         fprintf(stderr,
-                "[claw] CLAW_PREFIX too long (%zu bytes); ignoring and using defaults\n",
-                plen);
+                "[claw] CLAW_PREFIX too long (%llu bytes); ignoring and using defaults\n",
+                (unsigned long long)plen);
         return &_default_paths;
     }
     memcpy(prefix_buf, raw_prefix, plen + 1);
